Fix mapping of model name to use JSONAPISerializer.modelNameFromPayloadKey() method
Our team ran into a problem with a customized mapping for JSONAPI model names.  To fix this, we had to change ember-data-saverelationships to call the .modelNameFromPayloadKey() method which is the proper function for this.

https://api.emberjs.com/ember-data/3.15/classes/JSONAPISerializer/methods/modelNameFromPayloadKey?anchor=modelNameFromPayloadKey

NOTE: this change assumes the SaveRelationshipsMixin is meant to be used with JSONAPISerializer.extend (otherwise the method modelNameFromPayloadKey() wont exist.